helper: commands: extraids: allow to provide TV shows IMDB ID
When resolving episodes, sometimes the first aired date returned by
TheTVDB is not matching the one of IMDB. In such situations, a TV
shows might be discarded even though it was the right one. This
thus allows to instead match using the IMDB ID even when the year
does not match.

Fixes #98

Signed-off-by: Raphaël Beamonte <raphael.beamonte@gmail.com>